Further abstracted the service/model relationship for reset
